changed datatypes everywhere possible to 'usize'; tightened up 'fn new_cache' with vec assignement

+++ sim/src/main.rs
@@ -15,7 +15,7 @@ fn main() -> Result<(), Box<dyn Error>> {
 	let (set_bits, line_sum, block_bits, trace_file) = parse_them(&args)?;
 
 	// create a new instance of an empty cache with the specified number of sets and lines using the 'Cache' struct in the file: sim/src/lib.rs
-	let sets_sum = 2_u32.pow(set_bits as u32);
+	let sets_sum = (2_u32.pow(set_bits as u32)) as usize;
 	let mut new_cache = sim::Cache::new_cache(sets_sum, line_sum);
 
 	// read the trace file line by line; parse the operation, parse the set_index and tag from the address; update the cache
